// frontend/gestor_material/src/app/models/model.parsis.ts
export interface ModelParsis {
  id: number;
  razao: string;
  fantasia: string;
  cep: string;
  endereco: string;
  numero: string;
  complemento: string;
  bairro: string;
  cidade: string;
  uf: string;
  cnpj: string;
  ie: string;
  fone: string;
  email: string;
  ramoatividade: number;
  codigoibge: string;
  regime: number;
  ativo: string;
}